design: organization 메뉴 이름 변경 및 패딩 설정
