+++ app/src/main/java/com/easyhz/noffice/navigation/organization/OrganizationNavigation.kt
@@ -5,7 +5,7 @@ import androidx.navigation.NavController
 import androidx.navigation.NavGraphBuilder
 import androidx.navigation.NavOptions
 import androidx.navigation.compose.composable
-import com.easyhz.noffice.feature.organization.OrganizationScreen
+import com.easyhz.noffice.feature.organization.screen.organization.OrganizationScreen
 import com.easyhz.noffice.navigation.organization.screen.Organization
 
 internal fun NavGraphBuilder.organizationScreen(modifier: Modifier) {
