+++ feature/organization/src/main/java/com/easyhz/noffice/feature/organization/util/OrganizationTopBarMenu.kt
@@ -1,0 +1,10 @@
+package com.easyhz.noffice.feature.organization.util
+
+import com.easyhz.noffice.core.design_system.util.topBar.TopBarMenu
+
+enum class OrganizationTopBarMenu : TopBarMenu {
+    ORGANIZATION {
+        override val label: String
+            get() = "그룹"
+    }
+}
